Push a lot of the public functionality of the cart system into the ShoppingCart class itself
Methods in PaymentSystemService that take a shoppingcart as an argument
should just be in shoppingcart directly.
