+++ dspace/modules/payment-system/payment-webapp/src/main/java/org/dspace/app/xmlui/aspect/paymentsystem/ShoppingCartTransformer.java
@@ -98,7 +98,7 @@ public class ShoppingCartTransformer extends AbstractDSpaceTransformer {
             ShoppingCart shoppingCart = null;
             //create new transaction or update transaction id with item
             shoppingCart = paymentSystemService.getShoppingCartByItemId(context,item.getID());
-            paymentSystemService.updateTotal(context,shoppingCart);
+            shoppingCart.updateTotal(context);
 
             //add the order summary form (wrapped in div.ds-option-set for proper sidebar style)
