+++ dspace/modules/payment-system/payment-webapp/src/main/java/org/dspace/app/xmlui/aspect/shoppingcart/FlowShoppingcartUtils.java
@@ -113,7 +113,7 @@ public class FlowShoppingcartUtils {
                     }
                 
                 if (currency != null && !currency.equals(currencyOriginal)) {
-                    paymentSystemService.setCurrency(shoppingCart,currency);
+                    shoppingCart.setCurrency(currency);
                 }
                 else{
                     //only when the currency doesn't change then change the individual rate
@@ -144,7 +144,7 @@ public class FlowShoppingcartUtils {
                         }
                     }
                 
-                paymentSystemService.updateTotal(context,shoppingCart);
+                shoppingCart.updateTotal(context);
                 context.commit();
                 
                 result.setContinue(true);
